requirements.txt
openai-whisper
setuptools-rust
soundfile